% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{bundle}
\alias{bundle}
\title{Bundle packages installed in kaggle environment}
\usage{
bundle(output_dir = ".")
}
\arguments{
\item{output_dir}{Character with the output directory}
}
\description{
Bundle packages installed in kaggle environment
}
