% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{install}
\alias{install}
\title{Install packages in kaggle environment}
\usage{
install(include_data = FALSE)
}
\arguments{
\item{include_data}{Logical indicating if `sitsdata` must be installed}
}
\description{
Install packages in kaggle environment
}
