% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{patch}
\alias{patch}
\title{Patch kaggle environment with a bundle}
\usage{
patch(input_dir = NULL, lib_dir = .libPaths()[[1]], exclude = "torch")
}
\arguments{
\item{input_dir}{Character with the bundle directory. When `NULL`, the
bundle is searched in the kaggle input datasets}

\item{lib_dir}{Character with the library directory}

\item{exclude}{Character vector with packages to keep as they are}
}
\description{
Patch kaggle environment with a bundle
}
\note{
Use it before loading the packages of the bundle
}
