% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{install}
\alias{install}
\title{Install packages in kaggle environment}
\usage{
install(output_dir)
}
\arguments{
\item{output_dir}{Character with the output directory}
}
\description{
Install packages in kaggle environment
}
