% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{install}
\alias{install}
\title{Install packages in kaggle environment}
\usage{
install(...)
}
\arguments{
\item{...}{extra arguments for install packages}
}
\description{
Install packages in kaggle environment
}
